+++ LEGO1/lego/legoomni/include/legoutil.h
@@ -2,6 +2,9 @@
 #define LEGOUTIL_H
 
 #include "extra.h"
+#include "misc/legostorage.h"
+#include "misc/legotexture.h"
+#include "mxstring.h"
 #include "mxtypes.h"
 #include "mxutil.h"
 
@@ -10,6 +13,21 @@
 class MxAtomId;
 class LegoEntity;
 class LegoAnimPresenter;
+
+class LegoTexture;
+
+// SIZE 0x14
+class NamedTexture {
+public:
+	~NamedTexture() { delete m_texture; }
+	// FUNCTION: LEGO1 0x1003f920
+	const MxString* GetName() const { return &m_name; }
+	LegoTexture* GetTexture() { return m_texture; }
+
+private:
+	MxString m_name;        // 0x00
+	LegoTexture* m_texture; // 0x4
+};
 
 void FUN_1003e050(LegoAnimPresenter* p_presenter);
 Extra::ActionType MatchActionString(const char*);
@@ -20,5 +38,8 @@ void FUN_1003ef00(MxBool);
 void SetAppCursor(WPARAM p_wparam);
 MxBool FUN_1003ef60();
 MxBool RemoveFromWorld(MxAtomId& p_atomId1, MxS32 p_id1, MxAtomId& p_atomId2, MxS32 p_id2);
+NamedTexture* ReadNamedTexture(LegoFile* p_file);
+void FUN_1003f540(LegoFile* p_file, const char* p_filename);
+void WriteNamedTexture(LegoFile* p_file, NamedTexture* p_texture);
 
 #endif // LEGOUTIL_H
